mcp-local/requirements.txt
attrs==25.4.0
autobahn==19.11.2
automat==25.4.16
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
constantly==23.10.4
cryptography==46.0.5
h11==0.16.0
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
kiteconnect==5.0.1
langchain
langchain-community
outcome==1.3.0.post0
packaging==26.0
pip==23.0.1
plotly
pyasn1-modules==0.4.2
pyasn1==0.6.2
pycparser==3.0
pyopenssl==25.3.0
pyotp==2.9.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
scikit-learn
selenium==4.41.0
service-identity==24.2.0
setuptools==66.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio-websocket==0.12.2
trio==0.33.0
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wheel==0.38.4
wsproto==1.3.2
zope-interface==8.2
sqlalchemy
psycopg2-binary
statsmodels
pandas
numpy
yfinance
